refactored some logic
