+++ app/src/main/java/com/example/androidlearningapp/movies/data/api/MoviesRemoteSource.java
@@ -1,12 +1,10 @@
 package com.example.androidlearningapp.movies.data.api;
 
-import com.example.androidlearningapp.movies.data.listeners.GetMoviesListener;
+import com.example.androidlearningapp.movies.entity.MovieList;
+
+import io.reactivex.Observable;
 
 public interface MoviesRemoteSource {
 
-    void getMovies();
-
-    void setGetMoviesListener(GetMoviesListener getMoviesListener);
-
-    void dispose();
+    Observable<MovieList> getMovieListObservable();
 }
